Tweak English URLs to link to English links
We're changing the meaning of URLs - if the locale isn't given,
then the software uses heuristics and redirects to the best
available locale.  That means that if we want to click on a link
and be confident that we're going to an English locale version,
we have to specify the "en" locale in the link.

Note that although this will show as a "change" to all other
locales, we don't expect other locales to need to do anything,
as this is a change specifically for the English locale text.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>