Improve documentation, esp. cross-references
Add a number of hyperlinks between various important documents
in the docs/ directory.  These are added directly, not a
separate process, so that even users who view the files directly
can see the cross-references.

This adds a few other documentation tweaks, e.g., it adds
to the design document a citation to "You Are Not Google" at
<https://blog.bradfieldcs.com/you-are-not-google-84912cf44afb>.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>