Require user email uniqueness ONLY on local accounts
We previously checked email uniqueness in the user model, and
unintentionally required that the email be globally unique.

This commit instead requires case-insensitive uniqueness *only*
for local accounts, completely ignoring GitHub users (who are
handled differently).  In particular, we had a user who had
both her local account and GitHub account tied to the same
user email address; that should have worked, but didn't because
of the overly-restrictive check.

Note that this commit includes a regression test, one that did
*not* pass before but passes now.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>