Change "content_for" to "provide" (#1107)
Change from "content_for" to "provide", as it can be slightly faster
and more importantly enables potential future uses of streaming.

We use "content_for" in various views, but we *never* use its
ability to concatenate values.  Since we don't use that feature,
switch them all to "provide" instead, which can do the same thing
but doesn't have to search the whole template for values.
This is especially important if we ever use streaming in the future
for these cases.  Perhaps more importantly, we tend to create new things
based on existing constructs; if we use "provide" where practical,
future things will be more easily switched to streaming.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>