Add AppArmor to home page list of projects
AppArmor is a widely-used Linux Security Module (LSM) that provides
additional security by restricting capabilities using a per-program profile.
It is used by default in SuSE Linux Enterprise (SLES), openSUSE, and Ubuntu.
They have a badge, so I thought it'd be good to highlight it.
In particular, this is clearly a security-related project (we like to
highlight those), and it's different than the others (it's a module
of the Linux kernel, and we haven't listed those).

There are other LSMs, such as security-enhanced Linux (seLinux).
We're not taking a side, we're just noting that AppArmor is a widely-used
project that has earned a badge.

John Johansen and Emily Ratliff both said it was fine to add AppArmor
to the home page.  John Johansen uploaded the higher-resolution logo
used here (the logo file previously available on the AppArmor site was
a low-quality image).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>