Clarify: must EDIT project entries to change reminder status (#1124)
* Clarify: must EDIT project entries to change reminder status

We've received some emails that suggest that people are trying
to change the reminder status, but don't realize that they're
only in *show* mode and not *edit* mode.  Show and edit look
similar, but checkboxes in particular look nearly identical,
and that appears to be leading to the problem.

This commit addresses the problem by changing the text to try
to make it ABUNDANTLY clear that you have to EDIT the project
entry to change this value (just like any other value).

In the long term, we want to make it possible to link
*directly* to the edit mode for this value, to simplify this
further.  However, that requires non-trivial code changes,
and we would STILL want it to be clear to users anyway.
So let's make the situation clear with text right now,
because we *CAN* do that right now, without having to wait
for code changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Specifically note reminders on project entry screen

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>