Modify project_stats page to use ?type=uncommon (#1125)
This reduces delay and memory use, and fixes a bug.

Currently, when the project_stats page is shown, it's
supposed to show more graphs to admins.  This leads to a
rediculous amount of delay and memory use when the admin displays
the project_stats page, as well as some errors in display.

This commit instead shows the simple project_stats page in the
normal case, and shows *only* the less-common graphs if you include
"?type=uncommon" ("show the uncommon graphs").  This means that
you only see a particular set of graphs for any particular request.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>