Eliminate ActiveSupport.halt_callback_chains_on_return_false = true
Rails 5 changed how callback chains are interpreted. See:
https://blog.bigbinary.com/2016/02/13/
rails-5-does-not-halt-callback-chain-when-false-is-returned.html
In short, returning false DOES NOT stop the chain any more.
Rendering or redirecting still halts the chain (in development, this
produces a "Filter chain halted as .... rendered or redirected" message).
If you want to halt the chain, but do not render or redirect,
you can use this to halt the chain: throw(:abort)

This changes the code to throw an exception instead of just returning false,
and also renames several chain entries to make it clear that they
do NOT simply return a true/false value.
Thankfully, in almost all cases we were already invoking
render or redirect_to when we wanted to stop the chain.
Our negative testing also gives us confidence that the semantic change,
and this new code, isn't leading to a vulnerability.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>