Add details to license_location for CRAN. Fixes #1133
This commit clarifies criterion license_location's details, based on a
confusion in an R package. Below is the current state, a description of
what seems to be the confusion, and a proposed solution.

Criterion license_location has the following description:

> The project MUST post the license(s) of its results
> in a standard location in their source repository.

And originally just the following details:

> E.g., as a top-level file named LICENSE or COPYING.
> License filenames MAY be followed by an extension such
> as ".txt" or ".md".

However, an R project says it can't meet this because "R projects with
GPL-3 licenses are prohibited from including a separate license file,
as this is automatically provided by R itself."

However, I think this is a misunderstanding of the criterion. The
"Writing R Extensions" manual says that, "the “standard”
short specifications GPL-2 GPL-3 LGPL-2 LGPL-2.1 LGPL-3 AGPL-3
Artistic-2.0 BSD_2_clause BSD_3_clause MIT [are] made available via
https://www.R-project.org/Licenses/ and contained in subdirectory
share/licenses of the R source or home directory." Further, "Whereas
you should feel free to include a license file in your source
distribution, please do not arrange to install yet another copy
of the GNU COPYING or COPYING.LIB files but refer to the copies on
https://www.R-project.org/Licenses/ and included in the R distribution
(in directory share/licenses)."
Instead, these standard licenses given in the latter link are already
bundled by definition with R itself, and their inclusion is effectively
prohibited (a package won't be accepted on CRAN if it contains such
files). A similar thing happens when installing Debian packages.

However, a source repository (like GitHub) is not CRAN, so what we expect
is something like having a LICENSE in the repo and listing the LICENSE
in an exclusion file like .Rbuildignore.

This commit adds much more information to the details, including
specific examples, to counter this misunderstanding.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>