Remove Gemnasium.com references. Fixes #1132
Remove references to the Gemnasium badge and to the Gemnasium.com
site in the documentation as an actively-used system.
The Gemnasium.com service, as we used it, has closed.
I can imagine using the new service, but that would require other
changes (and would be handled separately).

For more information, see:
https://docs.gitlab.com/ee/user/project/import/gemnasium.html

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>