Add hypertext link for seeing user data in JSON format
The GDPR requires some sort of "portable format" for user data.
We have elected to provide this data in JSON format.
We've provided this data for some time, but the web page
didn't have a link to the JSON format.  It was *documented* how
to get this data in the API, but that's not the same thing.

This commit adds a hypertext link that makes it easy to see the
user data in JSON format.  Those who are not logged in, or are
logged in as a different non-logged-in user, will only receive the
*public* user data in JSON format.  This is just information such
as the public name, creation datetime, and last change datetime.
Logged-in users who look at their own data, and admins, will also see
in the JSON format the private data (which is currently the user email
address and preferred locale).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>