Add hypertext link for seeing user data in JSON format (#1137)
* Add hypertext link for seeing user data in JSON format

The GDPR requires some sort of "portable format" for user data.
We have elected to provide this data in JSON format.
We've provided this data for some time, but the web page
didn't have a link to the JSON format.  It was *documented* how
to get this data in the API, but that's not the same thing.

This commit adds a hypertext link that makes it easy to see the
user data in JSON format,  if the user is showing his own data OR
if the requesting user is an admin.  This probably isn't required
by the GDPR, but this seems like a nice touch.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>