Change user test fixture to use \n in blind index
Blind indexes must end in \n, but the default YAML
reader strips them out.  As a result, the email addresses
could not be found inside tests.

This commit replaces the user test fixture
with constants that end in \n.  These are precomputed
constants - it takes a moment to compute these, and
these get used often, so it makes sense to hand-calculate
them once and just store the results.  The comments in the
test fixture explain how to calculate more of them.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>