Encrypt email using attr_encrypted and blind_index gems (#1141, fixes #1138)
* Encrypt email using attr_encrypted and blind_index gems

This commit encrypts email using two gems,
attr_encrypted and blind_index, that are designed
specifically for this purpose.

* Change user test fixture to use \n in blind index

Blind indexes must end in \n, but the default YAML
reader strips them out.  As a result, the email addresses
could not be found inside tests.

This commit replaces the user test fixture
with constants that end in \n.  These are precomputed
constants - it takes a moment to compute these, and
these get used often, so it makes sense to hand-calculate
them once and just store the results.  The comments in the
test fixture explain how to calculate more of them.

* Remove direct SQL references to no-longer-present "email" field

* Tweak reminders_summary to decrypt email addresses

Simple database queries only return the *encrypted*
email addreses, so we need to decrypt separately.

* Fix implementation.md formatting

* User show: defensively program against bad email key

If the email decryption key is wrong, show an indicator
when doing "user show".  In normal execution this shouldn't
be an issue, but it's possible for this to be an issue when
downloading the database but not using the key.

* Modify user emailer to handle encrypted email addresses

Accessing email addresses is slightly different now because
they are encrypted; adjust the user emailer to handle that.

* Add some documentation about email encryption

In particular, add information about email encryption to the
the assurance case, and note the need for newlines in the
blind index (this isn't documented, and can cause trouble).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>