Gravatar (#1143)
* Only reveal gravatar MD5 hash if user has created it

We want to display gravatar images for local (custom) users,
since many people create them.  However, the URLs reveal the
MD5 of the user's downcased email address, and we want to
protect the email address.  If a user creates a gravatar,
then clearly the user has consented to that, but otherwise
it'd be nice to not reveal even a cryptographic hash of the
email address.

This commit adds a new boolean field "use_gravatar" to the
database "user" table.  If true on a local account, then we
will include on the user page a link to the user's gravatar
image (which uses the md5 of the email address); if not,
we'll use a bogus md5 so that we do not reveal anything
about the user's email address.

When a local account is created or edited, we check the
gravatar site to see if there's a gravar image or not, and
set/reset the use_gravatar value.

This also adds a new rake task "fix_use_gravatar"; just run
"rake fix_use_gravatar" to loop through all local accounts
and set their use_gravatar value appropriately.  We can
run this occasionally to make sure that it has the correct
value, though users can always edit their account to force
this anyway (so we don't really need to make a lot of changes).

This actually doesn't change much.  There are a lot of
changed lines, but almost all the changes are simply
the addition of VCR cassettes to record gravatar requests
in the test for creating or editing local accounts.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>