Remove field "unencrypted_email" from users table
This commit removes the field "unencrypted_email" from the users table,
along with associated indexes, as its own migration.

This commit is the final step in encrypting email addresses.
We did this in stages because email addresses are critical to
proper operation, so we wanted to make sure encryption worked before
we deleted this important data.  From this point on, email addresses
are encrypted within the database, providing some extra security
(particularly for the data-at-rest from the database).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>