Enable git login when the email keys aren't available (simpler way)
Enable git login when the email keys aren't available.

This makes it easier to test the system without making
the email keys available. That's a good thing, because
the more activities that don't require the keys, the
easier it is to protect those keys.

I had previously done this by rescuing the exception
raised if OpenSSL can't decrypt things, and then I had
to test that case, etc.

This commit instead does something much simpler:
instead of save!, just use update_columns.
Then there's no need to decrypt anything, so there's no error
to be triggered in the first place.  It also means that
we don't change the updated_at value of users merely when
they log in, which seems like a more appropriate use of
updated_at.  Now update_at is still changed when user-visible
data changes (such as the email address or locale preference),
but not for something as ephermeral as the last login time
(which is stored as a separate value anyway).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>