Add more fields to exported JSON user data
GDPR's "data portability" requirements require that we
provide users our data about them in a portable way.
We provide data in JSON format.  This commit adds a few
much-less useful fields, with the goal of providing
everything we reasonably can.  We don't want to share
digests, since in some circumstances those are like passwords
and they are randomly generated anyway.

That said, we can provide these, so let's do so:

*  json.activated
*  json.activated_at
*  json.reset_sent_at
*  json.last_login_at
*  json.use_gravatar

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>