Add rake task mass_email (e.g., for GDPR compliance)
Sometime we may need to contact users QUICKLY.  For example,
the GDPR requires user notification within 72 hours in some cases.

This commit adds a new rake task to do a mass email.
You set the subject, body, and condition via environment variables,
and then trigger the rake task.  This means that you do NOT
need to upload any code changes to send the mass email.

The implementation in this commit focuses on
being the "minimum viable product" that gets the job done.
This is NOT a user-visible function,
and changing it will only affect system administrators,
so changing the interface is no big deal.
We could add other features later, like HTML and internationalization.

This doesn't support HTML emails.  The point is to send a
relatively simple message quickly, so there's no need to complicate
things by including HTML.

This doesn't support internationalization.  In the expected use cases
we won't have time to get many or any translations done, and if we
do have some translations done, we can just include
ALL the available translations in a single body.

We hope that we will never need to use this task, but
it's best to be prepared.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>