Add a task to allow rekeying of email addresses
A critical property of keys and passwords, from a security viewpoint,
that it is *easy* to change them if they are revealed.
That doesn't work if there's no easy way to change them.

Here we provide a mechanism to rekey email addresses, if needed.
Most of the functionality is in the User model, so it's
a little easier to test.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>