Add more fields to exported JSON user data (#1154)
* Add more fields to exported JSON user data

GDPR's "data portability" requirements require that we
provide users our data about them in a portable way.
We provide data in JSON format.  This commit adds a few
much-less useful fields, with the goal of providing
everything we reasonably can.  We don't want to share
digests, since in some circumstances those are like passwords
and they are randomly generated anyway.

That said, we can provide these, so let's do so:

*  json.activated
*  json.activated_at
*  json.reset_sent_at
*  json.last_login_at
*  json.use_gravatar

* Add user.email_if_decryptable to simplify email access

Sometimes we want an email address, but only if it's
decryptable, otherwise just return a stub value.
Using this makes it easier to not use a key

* Modify user JSON to show more information

* Fix code formatting

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>