Implement BADGEAPP_DENY_LOGIN mode (#1157)
Implement the BADGEAPP_DENY_LOGIN mode.

If a non-blank value is set ("true" is recommended)
in the environmental variable BADGEAPP_DENY_LOGIN,
then no one can log in, no one can create a new account (sign up),
and no one can do anything that requires being logged in (users are always
treated as if they are not logged in).
This essentially prevents ANY changes by users (daily statistics
creates are unaffected).

From a security POV this is enforced by SessionsController#create (login),
UsersController#create (create new user/sign up), and
SessionsHelper#current_user (determine who current logged-in user is).
Some views disable the login and sign-in display, so that it's more
obvious to user what is going on.

This may be a useful mode to enable if there is a serious exploitable
security vulnerability that can only be exploited by users who are
logged in or can appear to log in.  Unlike *completely* disabling the
site, this mode allows people to see current information
(such as badge status, project data, and public user data).
Note that application admins cannot log in, or use their privileges,
when this mode is enabled.  Only hosting site admins can turn this mode
on or off (since they're the only ones who can set environment variables).

Before this commit, we could *completely* disable all access to the site
if there is a security issue, but that is not desirable.  In some
cases this mode is a reasonable alternative.  However, we have to
implement this mode *ahead* of time if we want to be able to use it
in the future.

There's not a lot of *main* code - a lot of the code here is to
implement tests of it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>