Fix error 404 (missing) response and its tests
The error 404 response wasn't working correctly.
However, our tests didn't notice, because in spite of the
problems the error page was correctly returning an error :-).
Fix the problems, and split it into two pickier tests
to prevent future regressions.

Originally this was tested by an "integration" test, but that
was misleading, as this operation is entirely within one controller.
So this commit moves the tests into the controller test, since that
is more consistent with how the rest of the tests are organized.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>