Change robots.txt implementation: Calculate + cache
To coordinate the fail2ban in rack_attack with robots.txt,
it's much easier to do some calculations in robots.txt.
That's fine, but we then need to cache robots.txt
(so calculations happen only once) and fix tests to deal with that.
This now does a fail2ban for a variety of paths, whether
or not they begin with a locale, and the robots.txt matches it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>