Adjust default ban list (no need for locales and enlarge)
Eliminate the complexity of per-locale bans.
We don't need to make the ban list per-locale, because
we'll halt processing before we redirect to a locale.

Expand the list of banned paths, to increase
the probability of detecting a script-kiddy-like probe.

This mechanism won't deter smart attackers, but squelching
naive attackers is useful.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>