Do not redirect if there is no such page and no locale
If there's no such page, and no locale in the URL,
we've been redirecting to a URL with a locale.
This wastes time in a round-trip and two calls for
no such page.  Instead, simply reply with an error page.

This results in less code and is faster.  It also eliminates
an bug where we redirect to "/" if given a URL
that looks like a locale ("/no-such-page").

This commit includes a number of extra tests to test various
circumstances. In particular, in error pages
we use the URL locale if given, and if there isn't one, we use
'Accept-Language' header provided by the browser to
determine the locale to use in an error page.

This also stops trying to figure out the locale for
/robots.txt.  We don't need that information at all, robots.txt
doesn't depend on the locale.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>