Improve assurance case documentation
Create new sections to specifically discuss access control
(identification, authentication, and authorization).
Make a few other improvements as well.

This now has a *few* evidence nodes (rectangles) on the figures in
the Claims, Arguments, and Evidence (CAE) notation.
I find it more convenient to discuss evidence in text than in
the figures in most cases, because editing figures takes more
work, but in a few cases it's worth putting them in the figure
(so let's do so).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>