Clarify 60-day window applies to *individual* projects, fixes #1159
Clarify (in "details") that the 60-day window in criterion
vulnerabilities_fixed_60_days only applies to the individual project.

Some projects are part of
larger umbrella organizations or larger projects, possibly in
multiple layers, and many projects feed their results to other
organizations and projects as part of a potentially-complex
supply chain. An individual project often cannot control
the rest, but an individual project can work to release
a vulnerability patch in a timely way.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>