Restructure assurance case to match ISO 12207; fixes #1184
Restructure the assurance case (doc/security.md) so that
many parts follow the structure of ISO/IEC/IEEE 12207
("Systems and software engineering - Software life cycle processes").
We consider every process, and include in the assurance case every
process important to it.

We don't claim that we conform to this standard, instead, we simply
use the 12207 structure to help ensure that we've considered
all of the lifecycle processes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>