Track test/vcr_cassettes/can_create_local_user.yml
This vcr_cassette was accidentally not checked in;
this commit adds the cassette to the repository.
This cassette is used in the local user creation test
so that we don't need to actually access the Gravatar page
"https://secure.gravatar.com/avatar/..." during a test.
We want to be able to run tests while disconnected from the
network for various reasons (e.g., so we can tests disconnected,
so that test conditions remain constant, and
so we don't unnecessarily burden external sites during tests).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>