In production, set config.assets.compile = false
In production, set config.assets.compile = false
as a minor speed and memory performance improvement.

This is recommended by Heroku's documentation at:
https://devcenter.heroku.com/articles/rails-asset-pipeline#
compile-set-to-true-in-production
which says: "If you have enabled your application to config.assets.compile =
true in production, your application might be very slow...
This setting is also known to cause other run-time instabilities and is
generally not recommended. Instead we recommend either precompiling all
of your assets on deploy (which is the default)..."

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>