rake translation:sync - Add Spanish translation of front page
This adds a Spanish translation to the key 3 main paragraphs
on the front page.  My thanks to Francisco Loaiza-Lemos
for doing this.  This means that someone who only speaks
Spanish can get a sense of what the project is about, and it
also means that when we demonstrate the different languages
on the front page it will actually look like we're
switching languages :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>