Identify many R tools and identify languages for some other tools
Criterion documentation_interface noted some programming languages
for some tools, but many other criteria didn't include that information,
and we didn't list many entries for R.
Expand on this information, so that developers can easily see
some tools for the programming languages their projects use
(either as things to consider or as examples they already know about).

We are *not* trying to create a complete list - we're simply providing
some information to help people.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>