Enable David A. Wheeler to get log issues from Google, fixes #1223
We need to receive Google notifications about problems with our site that
are generated by Google's webcrawler. In particular, I need to be able
to see them! To do this, we need to tweak our routes so that this URL:

https://bestpractices.coreinfrastructure.org/google75f94b1182a77eb8.html

Will generate this content exactly:

google-site-verification: google75f94b1182a77eb8.html

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>