Clarify that SPDX license expressions do not include quote marks
Make it clear in the section for entering SPDX license expressions
that single quotes and double quotes should not be included.
Note that in both the details and in the field title
(as a side-effect, internationalize the field title).

We hope that this will resolve this reported problem:
https://github.com/coreinfrastructure/best-practices-badge/issues/1226

That said, even if this doesn't resolve *that* problem, clarifying
what we need and internationalizing the title is a good idea.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>