Update gem rubyzip (1.2.1->1.2.2), resolves CVE-2018-1000544
This updates the gem rubyzip (1.2.1->1.2.2).
However, more importantly, this resolves CVE-2018-1000544.
Rubyzip had been very slow at fixing this, so we tracked it down and
determined that this vulnerability does not affect our system, so
we intentionally ignored its warnings.

Now that rubyzip has been fixed, this commit moves to the fixed version
AND removes all the "--ignore CVE-2018-1000544" options on
bundle audit check that had been necessary.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>