Relax implementation language verifier to fix #1235
Submitting https://github.com/dealii/dealii fails (per #1235)
because one of the languages reported is "Jupyter Notebook"
which has an embedded space.  Allow embedded spaces.

This tweak to the implementation_language verifier allows some
nonsense, but nothing serious, and we need to fix this quickly
so that a project can get started at all.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>