Fix format of "we have updated our requirements" text
We have a message saying
"We have updated our requirements for the criterion static_analysis",
but the Rails semantics for safe HTML in flash messages has changed.
As a result, its formatting is now wrong.

Instead of trying to significantly rework
app/controllers/projects_controller.rb, just remove the
not-really-necessary hypertext link.
This message only occurs in rare situations, and the hypertext link
isn't all that important.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>