Upgrade loofah->2.2.3 (CVE-2018-16468) fixes #1245
Upgrade loofah to version 2.2.3 to counter vulnerability CVE-2018-16468.
I've investigated this, and it does not appear that this
vulnerability in loofah (a component we use) is exploitable in our
configuration. Nevertheless, we try to update to not-vulnerable components
whenever practical, since sometimes attackers find clever ways to exploit
vulnerabilities in internal components.

Component report details: "In the Loofah gem for Ruby, through version 2.2.2,
unsanitized JavaScript may occur in sanitized output when a crafted SVG
element is republished. Users are advised to upgrade to version 2.2.3."
We don't allow users to submit elements to be republished as SVG,
and that's why this shouldn't be an issue.  That said, it might be
possible to somehow trick the system into doing it, so we'll update
to ensure there's no problem.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>