If additional rights list changes, invalidate cache. Fixes #1255
If someone edits additional rights, the changes don't
necessarily show on the project list, because we show the
old cached value instead.

This commit adds the list of additional rights to the cache key.
This ensures that when the list changes, the cache is invalidated.

If the performance is too slow, we could directly expire this
specific cache entry instead.  However, that adds a maintenance headache:
that would then create a subtle dependency on the cache key.
So let's try the "hard to get wrong" approach first, and switch
to the faster-performing version only if necessary.
I don't expect it to really impact performance, because for most
project this is an empty list.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>