Add newline in criteria, fixes #1244
This is a re-implementation of pull request #1244 by @marctjones per
https://github.com/coreinfrastructure/best-practices-badge/pull/1244

This commit adds a newline to separate the footer text from generated text.
This version, unlike the version from @marctjones, has a DCO signature
and modifies the source file doc/criteria-footer.markdown
(doc/criteria.md is a generated file).

My thanks to @marctjones for bringing this to our attention and
creating a first crack at it!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>