Remove unused translation keys (old "future" keys)
This commit eliminates a number of translations of keys that are now unused.

A change in translation.io's service has caused these
unused keys to be removed from the translations.  This is reasonable enough,
so accept this change (thus removing the obsolete translations).
It'd be possible to counteract this, but I can't see a great reason to bother.

Most of these keys are now unused due to
previous commit 41d7b9f5ebdec737b7bf7be73af45c188fd03685
(Tue Jun 27 08:55:07 2017 -0400), which removed "future" criteria from
the "passing" level.

This change looks confusing at first because these unused keys generally
*do* have a related key name that *is* in use.  For example, this change
removes in many locales this key:

> criteria.0.installation_common.description

but notice that this key remains:

> criteria.1.installation_common.description

What's more, because these are indented YAML files, all you immediately
see are "installation_common" / "description" as the keys.

Most of those "future" criteria became part of the
"silver" criteria level, so that's where they generally ended up.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>