Purge unused translation keys
Purge unused translation keys so that the translations on
translations.io do not include unused material.
This doesn't cause much of a change in the repo (we're removing
things we don't use anyway), but I thought it'd be important
to record that it happened.

This commit temporarily records what was purged in
,purgeable-translations.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>