Send email to old & new addresses if address changed on Rails 5.2.2
Rails 5.2.2 handles changes slightly different;
calculated fields (like email) are no longer reported as changes.
Since we store email as an encrypted value, this change
caused us to stop sending emails to both the old and new
email addresses.

We noticed this because of a drop in test coverage - we had 100%,
and then we didn't.  So look, 100% coverage actually had some value!

This change modifies cleanup_changes so that the "changes" we
use includes the email addresses (we have to handle them specially anyway).
It also hardens the test so that when we change the user email address
we require exactly one email to be sent to exactly the correct destinations.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>