Disable auto-paginate when getting repo list
This is an effort to fix issue #1268:
https://github.com/coreinfrastructure/best-practices-badge/issues/1268

It's failing to start a project on a user with a large
number of GitHub repos. If that is the cause of the problem, as seems
highly plausible, then limiting the number of repo records
we download from GitHub will resolve the problem.

Note: this commit does *not* fix the corresponding tests.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>