When getting user GitHub repo list, sort by pushed
When getting the user's list of GitHub repos, sort by the datetime
those repos were pushed to (so we skip old repos)
and limit loading to no more than MAX_GITHUB_REPOS_FROM_USER
(currently set to 50).

I've added a comment into the HTML (as an *HTML* comment)
so that someone who has questions about what is going on
will have multiple ways to find out.

We sort the repo list by full_name, which will make it
easier for a user to find a specific repo.

This commit will fail tests because I haven't reloaded
the VCR cassettes for testing.  I haven't bothered yet;
there's no point in doing that until we've agreed that we'll
do this.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>