Disable auto-paginate when getting repo list (#1270)
* Disable auto-paginate when getting repo list

This is an effort to fix issue #1268:
https://github.com/coreinfrastructure/best-practices-badge/issues/1268

It's failing to start a project on a user with a large
number of GitHub repos. If that is the cause of the problem, as seems
highly plausible, then limiting the number of repo records
we download from GitHub will resolve the problem.

Note: this commit does *not* fix the corresponding tests.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* When getting user GitHub repo list, sort by pushed

When getting the user's list of GitHub repos, sort by the datetime
those repos were pushed to (so we skip old repos)
and limit loading to no more than MAX_GITHUB_REPOS_FROM_USER
(currently set to 50).

I've added a comment into the HTML (as an *HTML* comment)
so that someone who has questions about what is going on
will have multiple ways to find out.

We sort the repo list by full_name, which will make it
easier for a user to find a specific repo.

This commit will fail tests because I haven't reloaded
the VCR cassettes for testing.  I haven't bothered yet;
there's no point in doing that until we've agreed that we'll
do this.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update test cassettes

We've changed how we show the early project repos
when creating a new badge.
This affects a number of tests, since a number of tests
check the interaction when creating a new project.
Update the VCR cassettes to handle that change.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Skip whitespace checks in cassettes

VCR cassettes are auto-generated, ignore their trailing whitespace.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update many VCR cassettes

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Modify VCR configuration so recording works

We now have varying per_page values.
Modify the VCR configuration so it stores things correctly.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>